add clean topo

+++ lab3/clean_topo.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$(id -u)" -ne 0 ]; then
+  echo "This script must be run as root." >&2
+  exit 1
+fi
+
 docker stop alice bob eve
 docker rm alice bob eve
